Currently, the Paxos Service trim logic is bounded by a max value (paxos_service_trim_max). This change dynamically modifies the max value when the number of logs to be trimmed is higher than paxos_service_trim_max.
The paxos_service_trim_max_multiplier has been added in case we want to increase paxos_service_trim_max by a certain factor when trim sizes are high. This limits the dynamically modified max value from growing beyond a certain limit so that the monitor is not overloaded with very large trims.

Signed-off-by: Aishwarya Mathuria <amathuri@redhat.com>